* debian:
  adding mangling to prune .dfsg (thanks lintian)
  Adding .dfsg since Documentation/ is pruned from source trees due to incomplete set of building materials to reproduce PDFs
  changelog entry for 1.9.2+svn680
  BP: jaunty/lenny -- no bdepends on gdcm2 for easy backporting

* dfsg: (149 commits)
  Pruning Documentation/ since it does not contain complete set of sources to build contained .pdfs
  sh => bash
  testing
  testing
  testing
  testing
  preparing for 1.9.2 release
  bug in previous incarnation of ROIStatistics
  bug in previous incarnation of ROIStatistics
  adding txt description to ROIStatistics
  BUG:  Needed the distance prior and not the spatial prior.  This bug doesn't effect initialization with prior probability images.
  ENH:  Added an option to specify the similarity metric radius uniquely in each direction, e.g. -m PR[fixed.nii.gz, moving.nii.gz, 1, 4x3x2]
  Minor changes to script, more needs to be done
  adding ROI mass and cluster csv output
  ENH: Added possibly different energy formulations.
  ENH:  Added possibly different energy formulations.
  BUG:  Added <algorithm> include for std::replace()
  ENH:  Reformulated the critical energy equation.
  ENH:  Reformulated the critical energy equation.
  update the shape update script to be consistent with btp
  ...

+++ Scripts/buildtemplateparallel.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
-VERSION="0.0.10"
+VERSION="0.0.13"
+
+# trap keyboard interrupt (control-c)
+trap control_c SIGINT
 
 # Uncomment the line below in case you have not set the ANTSPATH variable in your environment. 
-export ANTSPATH=${ANTSPATH:="$HOME/bin/ants/"}
+ export ANTSPATH=${ANTSPATH:="$HOME/bin/ants/"}
 
 #assuming .nii.gz as default file type. This is the case for ANTS 1.7 and up
 
@@ -12,9 +15,19 @@ function Usage {
 
 Usage: 
 
-$0 -d ImageDimension -o OUTPREFIX <other options> <images>
+`basename $0` -d ImageDimension -o OUTPREFIX <other options> <images>
 
-Compulsory arguments:
+Example Case:
+
+ echo " bash $0 -d 3 -m 30x50x20 -t GR  -s CC -c 1 -o MY -z InitialTemplate.nii.gz  *RF*T1x.nii.gz "
+ echo " in this case you use 30x50x20 iterations per registration "
+ echo " 4 iterations over template creation (that is the default) "
+ echo " with Greedy-SyN and CC metrics to guide the mapping. " 
+ echo " Output is prepended with MY and the initial template is InitialTemplate.nii.gz (optional). "
+ echo " the -c option is set to 1 which will try to use SGE to distribute the computation. "
+ echo " if you do not have SGE, use -c 0 or -c 2 --- read the help." 
+
+Compulsory arguments (minimal command line requires SGE cluster, otherwise use -c & -j options):
 
      -d:  ImageDimension: 2 or 3 (for 2 or 3 dimensional registration of single volume)
 	  ImageDimension: 4 (for 3 dimensional registration of time-series; requires FSL)
@@ -24,17 +37,18 @@ Compulsory arguments:
 <images>  List of images in the current directory, eg *_t1.nii.gz. Should be at the end
           of the command.
 
-NB: All files to be added to the template should be in the same directory.
+NB: All images to be added to the template should be in the same directory, and this script 
+should be invoked from that directory.
 
 Optional arguments:
 
-     -c:  Control for parallel computation (default 1) -- 0 == run serially,  1 == SGE qsub,  2 == use PEXEC
+     -c:  Control for parallel computation (default 1) -- 0 == run serially,  1 == SGE qsub,  2 == use PEXEC (localhost)
 
      -g:  Gradient step size (default 0.25) -- smaller in magnitude results in more cautious steps 
 
      -i:  Iteration limit (default 4) -- iterations of the template construction (Iteration limit)*NumImages registrations.
 
-     -j:  Number of cpu cores to use (default 2) -- requires "-c 2"
+     -j:  Number of cpu cores to use (default: 2; -- requires "-c 2"
 
      -m:  Max-iterations in each registration
 
@@ -48,7 +62,7 @@ Optional arguments:
      -t:  Type of transformation model used for registration. 
 
      -z:  Use this this volume as the target of all inputs. When not used, the script
-          will create an unbiased starting point by averaging all inputs.
+          will create an unbiased starting point by averaging all inputs. Use the full path!
 
 --------------------------------------------------------------------------------------
 ANTS was created by:
@@ -75,9 +89,9 @@ parallelize the registration of each subject to the template.
 
 Usage: 
 
-sh buildtemplateparallel.sh -d ImageDimension -o OUTPREFIX <other options> <images>
+`basename $0` -d ImageDimension -o OUTPREFIX <other options> <images>
 
-Compulsory arguments:
+Compulsory arguments (minimal command line requires SGE cluster, otherwise use -c & -j options)::
 
      -d:  ImageDimension: 2 or 3 (for 2 or 3 dimensional registration of single volume)
 	  ImageDimension: 4 (for 3 dimensional registration of time-series; requires FSL)
@@ -91,13 +105,17 @@ NB: All files to be added to the template should be in the same directory.
 
 Optional arguments:
 
-     -c:  Control for parallel computation   --- if set to zero, run serially, if set to 2 , use PEXEC , if set to 1 , use SGE qsub.  
+     -c:  Control for parallel computation   --- if set to zero, run serially, if set to 2 , use PEXEC (localhost), if set to 1 , use SGE qsub.  
 
      -g:  Gradient step size; smaller in magnitude results in more cautious steps (default 0.25)
 
      -i:  Iteration limit (default = 4) for template construction. requires 4*NumImages registrations.
 
-     -j:  Number of cpu cores to use (default 2)  --- set -c option to 2 to use this .
+     -j:  Number of cpu cores to use (default: 2; --- set -c option to 2 to use this .
+
+	  The optimal number of cpu cores to use for template generation depends on the availability of cores, the amount of 
+	  free working memory (RAM) and the resolution of the data. High resolution datasets typically require more RAM during 
+	  processing. Running out of RAM during a calculation will slow down all processing on your computer. 
 
      -m:  Max-iterations
 
@@ -143,7 +161,7 @@ Optional arguments:
              DD = Diffeomorphic Demons style exponential mapping 
 
      -z:  Use this this volume as the target of all inputs. When not used, the script
-          will create an unbiased starting point by averaging all inputs.
+          will create an unbiased starting point by averaging all inputs. Use the full path!
 
 Requirements:
 
@@ -518,6 +536,40 @@ function jobfnamepadding {
 
 }
 
+cleanup()
+# example cleanup function
+{
+  
+  cd ${currentdir}/
+
+  echo -en "\n*** Performing cleanup, please wait ***\n"
+
+# 1st attempt to kill all remaining processes
+# put all related processes in array
+  runningANTSpids=( `ps -C ANTS -C N3BiasFieldCorrection -C fslsplit | awk '{ printf "%s\n", $1 ; }'` )
+
+# debug only
+  #echo list 1: ${runningANTSpids[@]}
+
+# kill these processes, skip the first since it is text and not a PID  
+  for ((i = 1; i < ${#runningANTSpids[@]} ; i++)) 
+  do
+  echo "killing:  ${runningANTSpids[${i}]}"
+  kill ${runningANTSpids[${i}]} 
+  done
+  
+  return $?
+}
+ 
+control_c()
+# run if user hits control-c
+{
+  echo -en "\n*** User pressed CTRL + C ***\n"
+  cleanup
+  exit $?
+  echo -en "\n*** Script cancelled by user ***\n"
+}
+
 #initializing variables with global scope
 time_start=`date +%s`
 currentdir=`pwd`
@@ -525,11 +577,11 @@ nargs=$#
 
 MAXITERATIONS=30x90x20
 LABELIMAGE=0 # initialize optional parameter 
-METRICTYPE="PR" # initialize optional parameter
+METRICTYPE=CC # initialize optional parameter
 TRANSFORMATIONTYPE="GR" # initialize optional parameter
 N3CORRECT=1 # initialize optional parameter
 DOQSUB=1 # By default, buildtemplateparallel tries to do things in parallel
-GRADIENTSTEP="0.25" # Gradient step size, smaller in magnitude means more smaller (more cautious) steps
+GRADIENTSTEP=0.25 # Gradient step size, smaller in magnitude means more smaller (more cautious) steps
 ITERATIONLIMIT=4
 CORES=2
 TDIM=0
@@ -537,16 +589,20 @@ RIGID=0
 RIGIDTYPE=" --do-rigid" # set to an empty string to use affine initialization
 range=0
 REGTEMPLATE=target
+cpu_count=`cat /proc/cpuinfo | grep processor | wc -l`
 
-# Provide different output for Usage and Help 
+##Getting system info from linux can be done with these variables.
+# RAM=`cat /proc/meminfo | sed -n -e '/MemTotal/p' | awk '{ printf "%s %s\n", $2, $3 ; }' | cut -d " " -f 1`
+# RAMfree=`cat /proc/meminfo | sed -n -e '/MemFree/p' | awk '{ printf "%s %s\n", $2, $3 ; }' | cut -d " " -f 1`
+# cpu_free_ram=$((${RAMfree}/${cpu_count}))
+
+# Provide output for Help
 if [ "$1" == "-h" ]
     then
     Help >&2
-elif [ $nargs -lt 6 ]
-    then
-    Usage >&2
-fi
 
+fi
+OUTPUTNAME=antsBTP
 # reading command line arguments 
 while getopts "c:d:g:i:j:h:m:n:o:s:r:t:z:" OPT 
   do 
@@ -577,24 +633,6 @@ while getopts "c:d:g:i:j:h:m:n:o:s:r:t:z:" OPT
 	  ;;
       j) #number of cpu cores to use (default = 2)
 	  CORES=$OPTARG
-
-# will develop when I have time
-# This won't work on Mac
-#	testproc=`cat /proc/cpuinfo | grep processor | cut -d ':' -f 2 `
-# 	if [[ ${#testproc} -gt ${CORES} ]] ; 
-# 	then 
-# 		echo " You have spefied more threads than that there are cpu cores in your system. "
-# 		echo " This may slow down template generation"
-# 		echo " Maximum number of threads is: ${#testproc}"
-# 		echo " Press CTRL + C to exit and respecify "
-# 		sleep 5
-# 
-# 	elif [[ ${#testproc} -lt ${CORES} ]] ;
-# 	then
-# 		echo " You have spefied less threads than that there are cpu cores in your system. "
-# 		echo " This may slow down template generation"
-# 		echo " Maximum number of threads is: ${#testproc}"
-# 	fi  
 	  ;;
       m) #max iterations other than default 
 	  MAXITERATIONS=$OPTARG 
@@ -626,6 +664,27 @@ while getopts "c:d:g:i:j:h:m:n:o:s:r:t:z:" OPT
   esac
 done
 
+if [ $DOQSUB -eq 1 ] ; then
+  qq=`which  qsub`
+  if [  ${#qq} -lt 1 ] ; then 
+    echo do you have qsub?  if not, then choose another c option ... if so, then check where the qsub alias points ...  
+    exit 
+  fi
+fi 
+
+# Provide different output for Usage and Help 
+if [ ${TDIM} -eq 4 ] && [ $nargs -lt 5 ]
+    then
+    Usage >&2
+elif [ ${TDIM} -eq 4 ] && [ $nargs -eq 5 ]
+    then
+    echo ""
+    # This option is required to run 4D template creation on SGE with a minimal command line
+elif [ $nargs -lt 6 ]
+    then
+    Usage >&2
+fi
+
 #ANTSPATH=YOURANTSPATH
 if [  ${#ANTSPATH} -le 0 ]
     then
@@ -636,11 +695,11 @@ fi
 # Shiftsize is calculated because a variable amount of arguments can be used on the command line. 
 # The shiftsize variable will give the correct number of arguments to skip. Issuing shift $shiftsize will 
 # result in skipping that number of arguments on the command line, so that only the input images remain.
-shiftsize=`expr $OPTIND - 1`
+shiftsize=$(($OPTIND - 1))
 shift $shiftsize
 # The invocation of $* will now read all remaining arguments into the variable IMAGESETVARIABLE
 IMAGESETVARIABLE=$*
-NINFILES=`expr $nargs - $shiftsize`
+NINFILES=$(($nargs - $shiftsize))
 #test if FSL is available in case of 4D, exit if not
 if [  ${TDIM} -eq 4 ] && [  ${#FSLDIR} -le 0 ]
     then
@@ -650,20 +709,20 @@ fi
 if [ ${NINFILES} -eq 0 ]
     then
     echo "Please provide at least 2 filenames for the template."
-    echo "Use $0 -h for help"
+    echo "Use `basename $0` -h for help"
     exit 1
-elif [[ ${NINFILES} -eq 1 ]] && [[ -s fslnvols ]] 
+elif [[ ${NINFILES} -eq 1 ]] && [[ -s ${FSLDIR}/bin/fslnvols ]]
     then 
     range=`fslnvols ${IMAGESETVARIABLE}`
     if [ ${range} -eq 1 ] && [ ${TDIM} -ne 4 ]
 	then
 	echo "Please provide at least 2 filenames for the template."
-	echo "Use $0 -h for help"
+	echo "Use `basename $0` -h for help"
 	exit 1
     elif [ ${range} -gt 1 ] && [ ${TDIM} -ne 4 ]
 	then
 	echo "This is a multivolume file. Use -d 4"
-	echo "Use $0 -h for help"
+	echo "Use `basename $0` -h for help"
 	exit 1
     elif [ ${range} -gt 1 ] && [ ${TDIM} -eq 4 ]
 	then
@@ -729,7 +788,7 @@ for FLE in $ANTSSCRIPTNAME $PEXEC $SGE
   fi
 done
 
-
+# exit
 # check for an initial template image and perform rigid body registration if requested
 if [ ! -s $REGTEMPLATE ] 
     then 
@@ -737,20 +796,23 @@ if [ ! -s $REGTEMPLATE ]
     echo "--------------------------------------------------------------------------------------"
     echo " No initial template exists. Creating a population average image from the inputs."
     echo "--------------------------------------------------------------------------------------"
-    ${ANTSPATH}AverageImages $DIM populationmean.nii.gz 1 $IMAGESETVARIABLE
-    cp populationmean.nii.gz $TEMPLATE
-    
+    ${ANTSPATH}AverageImages $DIM $TEMPLATE 1 $IMAGESETVARIABLE  
 else
-    echo 
+    echo  
     echo "--------------------------------------------------------------------------------------"
-    echo " Initial template found.  This will be used for guiding the registration."
+    echo " Initial template found.  This will be used for guiding the registration. use : $REGTEMPLATE and $TEMPLATE "
     echo "--------------------------------------------------------------------------------------"
 	# now move the initial registration template to OUTPUTNAME, otherwise this input gets overwritten.
+     
     cp ${REGTEMPLATE} ${TEMPLATE}
     
 fi
 
 
+if [ ! -s $TEMPLATE ] ; then 
+  echo Your template : $TEMPLATE was not created.  This indicates trouble!  You may want to check correctness of your input parameters. exiting. 
+  exit 
+fi 
 # remove old job bash scripts 
 rm -f job*.sh 
 
@@ -832,7 +894,7 @@ if [ "$RIGID" -eq 1 ] ;
     # cleanup and save output in seperate folder
    
     mkdir rigid
-    mv rigid_*.nii.gz *Affine.txt rigid/
+    mv *.cfg rigid_*.nii.gz *Affine.txt rigid/
 
     # backup logs
     if [ $DOQSUB -eq 1 ]; 
@@ -932,8 +994,8 @@ while [  $i -lt ${ITERATIONLIMIT} ]
     elif [ $DOQSUB -eq 2 ] ; then 
 	echo $pexe 
 	echo $pexe >> job${count}_${i}.sh
-    elif  [ $DOQSUB -eq 0 ] ; then 
-	sh $exe
+    elif  [ $DOQSUB -eq 0 ] ; then
+	bash $exe
     fi
     
     # counter updated, but not directly used in this loop
@@ -994,7 +1056,7 @@ while [  $i -lt ${ITERATIONLIMIT} ]
   echo "--------------------------------------------------------------------------------------"		
 
   mkdir ${TRANSFORMATIONTYPE}_iteration_${i}
-  cp *${OUTPUTNAME}*.nii.gz ${TRANSFORMATIONTYPE}_iteration_${i}
+  cp ${TEMPLATENAME}warp*log.txt *.cfg *${OUTPUTNAME}*.nii.gz ${TRANSFORMATIONTYPE}_iteration_${i}/ 
 
   # backup logs
   if [ $DOQSUB -eq 1 ];
